+++ module/persistence/src/main/java/io/zhc1/realworld/persistence/ArticleCommentJpaRepositoryAdapter.java
@@ -3,6 +3,7 @@ package io.zhc1.realworld.persistence;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.context.annotation.Profile; // Added import
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,9 +13,10 @@ import io.zhc1.realworld.model.Article;
 import io.zhc1.realworld.model.ArticleComment;
 import io.zhc1.realworld.model.ArticleCommentRepository;
 
+@Profile("h2") // Added annotation
 @Repository
 @RequiredArgsConstructor
-class ArticleCommentRepositoryAdapter implements ArticleCommentRepository {
+class ArticleCommentJpaRepositoryAdapter implements ArticleCommentRepository { // Changed class name
     private final ArticleCommentJpaRepository articleCommentJpaRepository;
 
     @Override